Update README.md
Do
Bo — Does BO mean handsome?
In French Baby Names the meaning of the name Bo is: Respected; regarded highly (literal translation is Beautiful/handsome gaze). Also a Handsome.

Why do people call me BO?
What Does “Bo” Mean? This slang term is used to address someone who is of the male gender and/or a friend. It is thought to be a shortened or abbreviated version of the word “boy.”