Update task.md
What is Ö called in English?
The O umlaut is one of three German vowels that does not exist in the English alphabet. Ö does not have an equivalent sound in English. It's kind of like the sound you'd make when disgusted by something. Like all German vowels, the O umlaut has a long pronunciation and a short pronunciation.